% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runELA.R
\name{calcStability}
\alias{calcStability}
\title{Calculation stability indices.}
\usage{
calcStability(data = NULL, alpha, J, seitr = 1000, convTime = 10000)
}
\arguments{
\item{data}{: data is a matrix.}

\item{alpha}{: alpha is a vector which explicit/implicit preference. Use output of runSA function.}

\item{J}{: J is a matrix which shows species preference. Use output of runSA function.}
}
\description{
Claculating stable state energy, difference between sample's energy and stable state energy and stable state entropy.
}
