% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runELA.R
\name{ELA}
\alias{ELA}
\title{Energy landscape analysis}
\usage{
ELA(alpha = alpha, J = jj, SS.itr = 20000, FindingTip.itr = 10000)
}
\arguments{
\item{alpha}{: alpha is a vector which explicit/implicit preference. Use output of runSA function.}

\item{J}{: J is a matrix which shows species preference. Use output of runSA function.}

\item{SS.itr}{: SS.itr is a integer of iterator of stable state estimate.}

\item{FindingTip.itr}{: SS.itr is a integer of iterator of tipping points estimate.}
}
\description{
Estimating stable state and tipping point
}
