% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runELA.R
\name{ELAall}
\alias{ELAall}
\title{ELAall}
\usage{
ELAall(
  data = NULL,
  alpha,
  J,
  SS.itr = 20000,
  FindingTip.itr = 10000,
  plun = 0,
  seitr = 1000,
  convTime = 10000,
  thread = 1
)
}
\arguments{
\item{data}{: data is a binary matrix.}

\item{alpha}{: alpha is a vector which explicit/implicit preference. Use output of runSA function.}

\item{J}{: J is a matrix which shows species preference. Use output of runSA function.}
}
\description{
Run all ELA pipeline
}
