% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runELA.R
\name{ELplunning}
\alias{ELplunning}
\title{ELA Pluning}
\usage{
ELplunning(elasummary, th = 0.1)
}
\arguments{
\item{elasummary}{: elasummary is a output of ELA function.}

\item{th}{: th is a threshold which the range is 0 to 1.}
}
\description{
Pluning shallow basin
}
