\name{rela}

\alias{rela-package}

\alias{rela}

\docType{package}

\title{Scale item analysis with standard errors and principal axis factoring}

\description{The package's functions analyze scale items as a collective measure 
providing information both on the item and scale level. Cronbach's reliability estimates are
 reported both in the original and standardized form along with respective standard 
errors and confidence interval lower and upper bounds. Further the principal axis factoring 
function (paf) included in the package allows to preliminarily screen the items if they load
on the same single latent construct. Both functions' output can be viewed abbreviated using
 summary() as well as in graphical form using plot().}

\details{

\tabular{ll}{
Package: \tab rela\cr
Type: \tab Package\cr
Version: \tab 4.1\cr
Date: \tab 2009-10-25\cr
License: \tab Artistic\cr
LazyLoad: \tab yes\cr}

The package contains two function (itemanal and paf) which report a list of scale item analytic statistics.}


\author{Michael Chajewski ( \url{http://www.chajewski.com} )}

\references{
Cronbach, L. J. (1951). Coefficient alpha and the internal structure of tests. \emph{Psychometrika, 16(3)}, 297-334. \cr
\cr
Duhachek, A. & Iacobucci, D. (2004). Alpha's standard error (ASE): An accurate and precise confidence interval estimate. \emph{Journal of Applied Psychology, 89(5)}, 792-808. \cr
\cr
Kim, J., & Mueller, C. W. (1978). \emph{Introduction to factor analysis: What it is and how to do it}. SAGE Publications: Newbury Park, CA. \cr
\cr
Nunnally, J. C. & Bernstein, I. H. (1994). \emph{Psychometric theory} (3 ed.). McGraw-Hill: New York, NY. \cr
\cr
Kaiser, H. F. & Cerny, B. A. (1979). Factor analysis of the image correlation matrix. \emph{Educational and Psychological Measurement}, 39, 711-714.\cr
\cr
Pett, M. A., Lackey, N. R., & Sullivan, J. J. (2003). \emph{Making sense of factor analysis: The use of factor analysis for instrument development in health care research}. SAGE Publications: Thousand Oaks, CA.}

