% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSA.R
\name{runSA}
\alias{runSA}
\title{Stochastic Approximation}
\usage{
runSA(data = NULL, env = NULL, rep = 16, max.itr = 10000)
}
\arguments{
\item{data}{: data is a matrix. Row is sample, column is species.}

\item{env}{: env is a matrix. You should standardize it.}

\item{rep}{; rep is a integer.}

\item{max.itr}{: max.itr is a integer.}
}
\description{
Estimating species interaction, environemnt preference.
}
