% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSA.R
\name{runSAparallel}
\alias{runSAparallel}
\title{Parallel stochastic Approximation}
\usage{
runSAparallel(data = NULL, env = NULL, rep = 16, max.itr = 10000, thread = 1)
}
\description{
Parallel stochastic Approximation
}
